# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AUTH API",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Creat comment",
                "description": "Creat comment",
                "operationId": "874f1f8b00690d0329ee4ecc31573ae7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text",
                                    "posts_id"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Первый Комментарий"
                                    },
                                    "posts_id": {
                                        "type": "number",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "posts_id": {
                                            "type": "string",
                                            "example": "15"
                                        },
                                        "text": {
                                            "type": "string",
                                            "example": "Первый Комментарий"
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "id": {
                                            "type": "number",
                                            "example": 12
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "text": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Поле 'text' обязательное!"
                                                    }
                                                },
                                                "posts_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Поле 'posts_id' обязательное!"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        },
                                        "Description": {
                                            "type": "string",
                                            "example": "Missing or Invalid Access Token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/comments/{id}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Edit comment",
                "description": "Edit comment",
                "operationId": "b3af228ff168b3937518a458325dc913",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Первый Комментарий NEW"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "posts_id": {
                                            "type": "string",
                                            "example": "15"
                                        },
                                        "text": {
                                            "type": "string",
                                            "example": "Первый Комментарий NEW"
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "id": {
                                            "type": "number",
                                            "example": 12
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "text": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Поле 'text' обязательное!"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        },
                                        "Description": {
                                            "type": "string",
                                            "example": "Missing or Invalid Access Token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Remove comment",
                "description": "Remove comment",
                "operationId": "7dcaf275761132541f46b6a85ad48b80",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "succses"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        },
                                        "Description": {
                                            "type": "string",
                                            "example": "Missing or Invalid Access Token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}